+++ js/constants.js
@@ -1,0 +1,25 @@
+const OFFERS_COUNT = 10;
+
+const LATITUDE = {
+  MIN: 35.65,
+  MAX: 35.7,
+};
+const LONGITUDE = {
+  MIN: 139.7,
+  MAX: 139.8,
+};
+const OFFER_TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+
+const TIME = {
+  CHECKIN: ['12:00', '13:00', '14:00'],
+  CHECKOUT: ['12:00', '13:00', '14:00'],
+};
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export { OFFERS_COUNT, LATITUDE, LONGITUDE, OFFER_TYPES, TIME, FEATURES, PHOTOS };
